% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LH_plot.R
\name{LH_plot}
\alias{LH_plot}
\title{LH_plot}
\usage{
LH_plot <- function(sample_output, output_type = 'none')
}
\arguments{
\item{sample_output}{Output from LH_sample()}

\item{output_type}{How plots are written and saved: ‘none’ displays in R graphics device only, ’pdf’ produces a single .pdf with all plots, and ‘png' produces a separate .png for each plot.}
}
\description{
A function that produces plots from the output of LH_sample().
}
\keyword{plot}
